# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Jobify",
        "description": "Thông tin dự án Jobify",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8040/api",
            "description": "Local server"
        }
    ],
    "paths": {
        "/auth/login": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Process Login User",
                "description": "Log in to your account and return the JWT token.",
                "operationId": "4a0653990bae153573fca45c02b86632",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "example": "admin@example.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "Admin@12"
                                    },
                                    "remember": {
                                        "type": "boolean",
                                        "example": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "User login successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "User login successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Wrong account or password",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Wrong account or password"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/auth/logout": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Logout user",
                "description": "Log out a user by invalidating the JWT token.",
                "operationId": "615f1ac2443597e75609845bf0366ab4",
                "responses": {
                    "200": {
                        "description": "Logout successful",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Người dùng đã đăng xuất"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "The user is not logged in",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "The user is not logged in"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Server Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "User logout failure"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/auth/job-seeker-register": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Create New JobSeeker Account",
                "operationId": "resigterJobSeeker",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "full_name",
                                    "email",
                                    "password",
                                    "password_confirmation",
                                    "phone_number"
                                ],
                                "properties": {
                                    "full_name": {
                                        "description": "Họ tên người tìm việc",
                                        "type": "string",
                                        "example": "Người tìm việc 1"
                                    },
                                    "email": {
                                        "description": "Email người tìm việc",
                                        "type": "string",
                                        "example": "jobseeker123@gmail.com"
                                    },
                                    "password": {
                                        "description": "Mật khẩu",
                                        "type": "string",
                                        "example": "Timviec@123"
                                    },
                                    "password_confirmation": {
                                        "description": "Nhập lại mật khẩu",
                                        "type": "string",
                                        "example": "Timviec@123"
                                    },
                                    "phone_number": {
                                        "description": "Nhập số điện thoại vào nhé",
                                        "type": "string",
                                        "example": "0912345678"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Registration successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Registration successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Registration error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Registration error"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "JWT Bearer token",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}